refactor(gateway, shared, security): Unify ErrorResponse and ApiError into shared ApiError
