+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/app/api/ApiResponse.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 import org.springframework.http.HttpStatus;
+import ru.pavbatol.myplace.shared.dto.api.ApiError;
 
 @Data
 @Accessors(chain = true)
